Room/Hero inventory, Hero use item, and displaydungeon aorund hero
added the functionality of both the hero and dungeon room invenotry such as adding removing and displaying the inventories for both. Made it so if a item is usable you can use the item such as heal. Also added trap item for rooms to have. Created a method that displays the dungeon near the Character.